Fix margin in sign in / sign up forms divider
When the "Or fill the following form" text was translated to another
language or customized by administrators, the text could span over two
lines. Since the element had a fixed height, it could overlap with the
text below it.

So now we're using an element with a relative position to achieve the
same effect (have the contents of the elements on top of its border).